requirements.txt
anthropic>=0.40.0
google-auth>=2.36.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.155.0
python-dotenv>=1.0.0
requests>=2.32.0
